chore(web): modularizes app/browser's touch-based language menu
