+++ web/src/app/browser/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { _CreateElement } from './createElement.js';
+export { getStyleValue } from './getStyleValue.js';
+export { getViewportScale } from './getViewportScale.js';
